predict变更

+++ utils/utils.py
@@ -92,7 +92,7 @@ def KNN_by_iter(Feature_train, target_train, Feature_test, target_test, k, devic
         for item in range(Feature_test.shape[0]):
             K = copy.copy(k)
             while True:
-                target_train_index = target_train[Index[item, :K].astype('int32'), 0]
+                target_train_index = target_train[Index[item, :K], 0]
                 print(target_train_index)
                 if len(np.unique(target_train_index)) >= 5:
                     break
